+++ src/routes/v1/index.js
@@ -7,17 +7,11 @@ import identityRouter from './identity';
 import utilityRouter from './utility';
 import dpkiRouter from './dpki';
 
-import * as config from '../../config';
-
 const router = express.Router();
 
-if (config.role === 'rp') {
-  router.use('/rp', rpRouter);
-} else if (config.role === 'idp') {
-  router.use('/idp', idpRouter);
-} else if (config.role === 'as') {
-  router.use('/as', asRouter);
-}
+router.use('/rp', rpRouter);
+router.use('/idp', idpRouter);
+router.use('/as', asRouter);
 router.use('/identity', identityRouter);
 router.use('/utility', utilityRouter);
 router.use('/dpki', dpkiRouter);